Use optional properties in Exp to save space
This is an optimisation motivated by:

1. Large database files

Most subexpressions had 6 unused properties. This scales badly with
large expressions. It creates large files which are unwieldy to view
and edit.

2. Slow evaluation

I suspected that excess memory usage was contributing to long
evaluation times. It seems like this was true and evaluation is faster
now, although my testing was not very rigorous.