Implement new pattern matching system
This introduces binds which are distinct from variables. For example:

*x -> x

The identity function binds x and returns the variable x. Evaluation
of applying a symbol:

(*x -> x) :a
*x = :a in x
:a

Patterns are evaluated. This allows you to abstract over patterns:

*pair = *a, *b in
*swap = pair -> b, a in
swap (:a, :b)

Which evaluates like this (steps omitted):

*pair = *a, *b in
*swap = pair -> b, a in
swap (:a, :b)

*pair = *a, *b in
(pair -> b, a) (:a, :b)

(*a, *b -> b, a) (:a, :b)

:b, :a

Notice that by using the `pair` pattern, we introduced *a and *b.

The result of evaluation is always "data". This means you can only
ever pattern match on data. See the `Data` type in exp.ts