Use inline-block instead of flex
The main issue with flex is that the flex container width is
calculated based on total width of the items, even if those items are
arranged vertically - making the container unnecessarily wide.

Flex is still useful for keeping the left hand side of the exp grouped
with the operator. However this requires an additional inline-block
container since a container can't just be "flex on the inside".