Implement new unary exp UI behaviour
Unary expressions are now borderless textareas. Symbols are dark and
bold. Binds are light and italic.

Clicking on unary exps (textareas) just focuses them (exp focus, not a
focus event). This is consistent with other exps.

When inputting, you get default mouse behaviour, so you can select
text etc.

When a unary exp is hidden, it appears as an ellipsis, unless you're
inputting.