Fix variable shadowing for let
Although now I realise it doesn't quite make sense. What if there's an
unevaluated expression that will evaluate to a shadowing bind?

For example, this evaluates to :b, because shadowing occurs:

*x = :a in
*x = :b in
x

Whereas you get :a here:

*shadower = (*x = :b) in
*x = :a in
shadower in
x

Because it evaluates to this:

*x = :a in
*shadower = (*x = :b) in
*x = :b in
x

It evaluates to that because `shadower` could be recursive. When we
perform substitution, we also substitute the original binding in order
to support recursion.

The semantics here are obviously very messy and will need to change.